+++ tests/actors/state_observer.py
@@ -8,7 +8,6 @@ import HABApp.rule.rule
 import habapp_rules.actors.state_observer
 import habapp_rules.core.state_machine_rule
 import tests.helper.oh_item
-import tests.helper.rule_runner
 import tests.helper.test_case_base
 import tests.helper.timer
 
@@ -396,3 +395,73 @@ class TestStateObserverRollerShutter(tests.helper.test_case_base.TestCaseBase):
 		tests.helper.oh_item.item_state_change_event("Unittest_RollerShutter", 60)
 		self.assertEqual(60, self._observer_jalousie.value)
 		self._cb_manual.assert_called_once_with(unittest.mock.ANY)
+
+
+class TestStateObserverNumber(tests.helper.test_case_base.TestCaseBase):
+	"""Tests cases for testing StateObserver for number item."""
+
+	def setUp(self) -> None:
+		"""Setup test case."""
+		tests.helper.test_case_base.TestCaseBase.setUp(self)
+
+		tests.helper.oh_item.add_mock_item(HABApp.openhab.items.NumberItem, "Unittest_Number", None)
+
+		self._cb_manual = unittest.mock.MagicMock()
+		self._observer_switch = habapp_rules.actors.state_observer.StateObserverNumber("Unittest_Number", self._cb_manual)
+
+	def test_command_from_habapp(self):
+		"""Test HABApp rule triggers a command -> no manual should be detected."""
+		for value in [0, 0, 42, 42, 0]:
+			self._observer_switch.send_command(value)
+			tests.helper.oh_item.item_command_event("Unittest_Number", value)
+			tests.helper.oh_item.item_state_change_event("Unittest_Number", value)
+			self._cb_manual.assert_not_called()
+
+	def test_manu_from_openhab(self):
+		"""Test manual detection from openHAB."""
+		TestCase = collections.namedtuple("TestCase", "command, cb_manual_called")
+
+		test_cases = [
+			TestCase(0, False),
+			TestCase(42, True),
+			TestCase(0, True),
+			TestCase(0, False),
+			TestCase(1000, True),
+		]
+
+		for test_case in test_cases:
+			self._cb_manual.reset_mock()
+
+			tests.helper.oh_item.item_state_change_event("Unittest_Number", test_case.command)
+
+			self.assertEqual(test_case.cb_manual_called, self._cb_manual.called)
+			if test_case.cb_manual_called:
+				self._cb_manual.assert_called_with(unittest.mock.ANY)
+
+	def test_manu_from_extern(self):
+		"""Test manual detection from extern."""
+		TestCase = collections.namedtuple("TestCase", "command, cb_manual_called")
+
+		test_cases = [
+			TestCase(0, False),
+			TestCase(42, True),
+			TestCase(0, True),
+			TestCase(0, False),
+			TestCase(1000, True),
+		]
+
+		for test_case in test_cases:
+			self._cb_manual.reset_mock()
+
+			tests.helper.oh_item.item_state_change_event("Unittest_Number", test_case.command)
+
+			self.assertEqual(test_case.cb_manual_called, self._cb_manual.called)
+			if test_case.cb_manual_called:
+				self._cb_manual.assert_called_with(unittest.mock.ANY)
+			tests.helper.oh_item.item_state_change_event("Unittest_Number", test_case.command)
+			self.assertEqual(test_case.command, self._observer_switch.value)
+
+	def test_send_command_exception(self):
+		"""Test if correct exceptions is raised."""
+		with self.assertRaises(ValueError):
+			self._observer_switch.send_command("ON")