+++ tests/sensors/sun.py
@@ -7,7 +7,6 @@ import HABApp
 import habapp_rules.sensors.sun
 import tests.helper.graph_machines
 import tests.helper.oh_item
-import tests.helper.rule_runner
 import tests.helper.test_case_base